added carts and make orders

+++ app/src/main/java/com/vkr/vkrmobile/model/data/net/response/cart/CartItemResponse.kt
@@ -5,5 +5,5 @@ import com.vkr.vkrmobile.model.data.net.response.product.ProductResponse
 
 class CartItemResponse (
     @SerializedName("productData") val productData: ProductResponse,
-    @SerializedName("count") val count: Long
+    @SerializedName("count") val count: Int
 )
